refactor: refactor code layout by moving ui to another crates
