+++ src/ui/app/mod.rs
@@ -1,0 +1,3 @@
+pub mod anime;
+pub mod app;
+pub mod ln;
