+++ src/ui/input.rs
@@ -1,0 +1,4 @@
+pub enum InputMode {
+    Normal,
+    Editing,
+}
